hw11 добавление интерфейса для ATM;      выделение отдельного файла под enum для номиналов;      переработка конструктора класса ATM
